+++ c/baokeshed_avx512.c
@@ -38,14 +38,14 @@ INLINE __m256i xor_256(__m256i a, __m256i b) { return _mm256_xor_si256(a, b); }
 
 INLINE __m512i xor_512(__m512i a, __m512i b) { return _mm512_xor_si512(a, b); }
 
-INLINE __m128i set1_128(uint32_t x) { return _mm_set1_epi32(x); }
+INLINE __m128i set1_128(uint32_t x) { return _mm_set1_epi32((int32_t)x); }
 
-INLINE __m256i set1_256(uint32_t x) { return _mm256_set1_epi32(x); }
+INLINE __m256i set1_256(uint32_t x) { return _mm256_set1_epi32((int32_t)x); }
 
-INLINE __m512i set1_512(uint32_t x) { return _mm512_set1_epi32(x); }
+INLINE __m512i set1_512(uint32_t x) { return _mm512_set1_epi32((int32_t)x); }
 
 INLINE __m128i set4(uint32_t a, uint32_t b, uint32_t c, uint32_t d) {
-  return _mm_setr_epi32(a, b, c, d);
+  return _mm_setr_epi32((int32_t)a, (int32_t)b, (int32_t)c, (int32_t)d);
 }
 
 INLINE __m128i rot16_128(__m128i x) { return _mm_ror_epi32(x, 16); }
@@ -456,7 +456,7 @@ INLINE void transpose_msg_vecs4(const uint8_t *const *inputs,
 
 INLINE void load_offsets4(uint64_t offset, const uint64_t deltas[4],
                           __m128i *out_lo, __m128i *out_hi) {
-  __m256i a = _mm256_add_epi64(_mm256_set1_epi64x(offset),
+  __m256i a = _mm256_add_epi64(_mm256_set1_epi64x((int64_t)offset),
                                _mm256_loadu_si256((const __m256i *)deltas));
   *out_lo = _mm256_cvtepi64_epi32(a);
   *out_hi = _mm256_cvtepi64_epi32(_mm256_srli_epi64(a, 32));
@@ -720,7 +720,7 @@ INLINE void transpose_msg_vecs8(const uint8_t *const *inputs,
 
 INLINE void load_offsets8(uint64_t offset, const uint64_t deltas[8],
                           __m256i *out_lo, __m256i *out_hi) {
-  __m512i a = _mm512_add_epi64(_mm512_set1_epi64(offset),
+  __m512i a = _mm512_add_epi64(_mm512_set1_epi64((int64_t)offset),
                                _mm512_loadu_si512((const __m512i *)deltas));
   *out_lo = _mm512_cvtepi64_epi32(a);
   *out_hi = _mm512_cvtepi64_epi32(_mm512_srli_epi64(a, 32));
@@ -1042,9 +1042,9 @@ INLINE void transpose_msg_vecs16(const uint8_t *const *inputs,
 
 INLINE void load_offsets16(uint64_t offset, const uint64_t deltas[16],
                            __m512i *out_lo, __m512i *out_hi) {
-  __m512i a = _mm512_add_epi64(_mm512_set1_epi64(offset),
+  __m512i a = _mm512_add_epi64(_mm512_set1_epi64((int64_t)offset),
                                _mm512_loadu_si512((const __m512i *)&deltas[0]));
-  __m512i b = _mm512_add_epi64(_mm512_set1_epi64(offset),
+  __m512i b = _mm512_add_epi64(_mm512_set1_epi64((int64_t)offset),
                                _mm512_loadu_si512((const __m512i *)&deltas[8]));
   __m256i a_lo = _mm512_cvtepi64_epi32(a);
   __m256i b_lo = _mm512_cvtepi64_epi32(b);
